Editor: Use the Editor component in edit-post
