+++ packages/editor/src/private-apis.native.js
@@ -7,7 +7,6 @@ import * as interfaceApis from '@wordpress/interface';
  * Internal dependencies
  */
 import VisualEditor from './components/visual-editor';
-import { ExperimentalEditorProvider } from './components/provider';
 import { lock } from './lock-unlock';
 import { EntitiesSavedStatesExtensible } from './components/entities-saved-states';
 import useBlockEditorSettings from './components/provider/use-block-editor-settings';
@@ -22,7 +21,6 @@ const { store: interfaceStore, ...remainingInterfaceApis } = interfaceApis;
 export const privateApis = {};
 lock( privateApis, {
 	VisualEditor,
-	ExperimentalEditorProvider,
 	EntitiesSavedStatesExtensible,
 	PluginPostExcerpt,
 	PreferencesModal,
